# Use a default vocabulary ID when guessing vocabulary, to save lookups