Part of #836

This change makes better use of the `defaultVocabId` parameter in `guessVocabularyFromURI` method in two places. When handling search results and concept properties, the ID of the current vocabulary is used as default so that when a URI that is within the URI space of current vocabulary (e.g. YSO) is encountered, it is assumed to be within the current vocabulary, even though theoretically it could be from another vocabulary (e.g. YSO Places) which uses the same URI namespace.

This may be a little bit risky, for example if a YSO concept links to a YSO Places concept using a property that is not a mapping property.