@osma 2019-04-23
Tried to fix the merge conflict using the GitHub UI but it didn't seem to work. Trying again locally...