@osma 2019-04-23
Fixed the conflict locally, rebased over current master, squashed the two commits into one.