@osma 2019-04-23
The effect should be most noticeable for queries in YSO (or YSO places) that return a lot of results. For example short strings like "aut" or "kar". In those cases currently Skosmos has to do a bit of extra work to match each concept to a vocabulary. This patch should reduce that work and thus improve the response time for generating the search result page.

Will test this soon...