@osma 2019-04-23
YSO (Finnish) search for "aut" (89 hits) goes from 2.4 seconds to 1.0 seconds.
YSO (Finnish) search for "kar" (146 hits) goes from 2.8 seconds to 1.1 seconds.
YSO (Finnish) search for "aamu" (4 hits) goes from 0.49 seconds to 0.38 seconds.
YSO concept page view goes from 0.21 seconds to 0.19 seconds.

As speculated above, the speedup is most noticeable for long search result lists.

The one problematic case I can think of is YSO concepts that link to YSO places, for example [Johanneksenkirkko (Helsinki)](http://www.yso.fi/onto/yso/p25064). Currently the link is to Helsinki in YSA but I suspect it will be changed soon to point to YSO Places as YSA is being deprecated. I will have to test this separately before merging...