@osma 2019-04-23
It turns out the display of (non-mapping) links to YSO Places concepts on YSO concept pages is already broken:

![image](https://user-images.githubusercontent.com/1132830/56584058-9a2af980-65e3-11e9-8316-27f5274972a0.png)

(The place is shown as `ysopaikat:p94137` instead of its label Helsinki)

This PR doesn't change the situation.