@osma 2019-04-23
Reported the concept page issue as #868.